Use Interface to specify the common interface between controllers
